refactor: moved import userAssessmentEntity from assessment to user-assessment
